requirements.txt
Flask
flask-pymongo
pandas
scikit-learn
matplotlib
seaborn
python-dotenv
